+++ packages/CollapsibleCard/src/CollapsibleCard.styles.js
@@ -43,6 +43,7 @@ function getBorder(isEditing, position) {
 
 export const CollapsibleCard = styled.div(
   ({ isEditing, isCollapsed, position }) => css`
+    background: ${tokens.color.white};
     border: ${getBorder(isEditing, position)};
     border-radius: ${getBorderRadius(position, isCollapsed, false)};
     box-shadow: ${getBoxShadow(isEditing, position)};