+++ packages/InlineEditing/stories/examples/InlineInput.simple.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Input from "@paprika/input";
-import Editor from "../../src/Editor";
 import Card from "@paprika/card";
+import Editor from "../../src/Editor";
 
 function InlineInput(props) {
   const { value, onSubmit, isEditing, onClick } = props;