+++ packages/DataHeader/CHANGELOG.md
@@ -1,3 +1,39 @@
 <!-- content -->
 
+## 1.0.12-next.0
+
+### Patch Changes
+
+- Updated dependencies [c265721]
+  - @paprika/popover@1.2.0-next.0
+  - @paprika/data-field@1.0.12-next.0
+  - @paprika/overflow-menu@1.0.14-next.0
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [12938b6]
+- Updated dependencies [0b3b614]
+  - @paprika/popover@1.1.0
+  - @paprika/toast@1.1.0
+  - @paprika/data-field@1.0.11
+  - @paprika/overflow-menu@1.0.13
+
+## 1.0.11-next.1
+
+### Patch Changes
+
+- Updated dependencies [0b3b614]
+  - @paprika/toast@1.1.0-next.0
+
+## 1.0.11-next.0
+
+### Patch Changes
+
+- Updated dependencies [12938b6]
+  - @paprika/popover@1.1.0-next.0
+  - @paprika/data-field@1.0.11-next.0
+  - @paprika/overflow-menu@1.0.13-next.0
+
 <!-- eoContent -->