+++ packages/ListBoxBrowser/CHANGELOG.md
@@ -1,5 +1,50 @@
 # Changelog
 
+## 1.0.17-next.0
+
+### Patch Changes
+
+- @paprika/list-box@3.0.2-next.0
+
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies [0b3b614]
+  - @paprika/list-box@3.0.1
+
+## 1.0.16-next.0
+
+### Patch Changes
+
+- @paprika/list-box@3.0.1-next.0
+
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [47607db]
+  - @paprika/list-box@3.0.0
+
+## 1.0.15-next.0
+
+### Patch Changes
+
+- Updated dependencies [47607db]
+  - @paprika/list-box@3.0.0-next.0
+
+## 1.0.14
+
+### Patch Changes
+
+- da9bc25: Fixed column styling, adjusted alignment of checkboxes/radios
+
+## 1.0.14-next.0
+
+### Patch Changes
+
+- da9bc25: Fixed column styling, adjusted alignment of checkboxes/radios
+
 ## 1.0.13
 
 ### Patch Changes