+++ packages/OverflowMenu/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 1.0.14-next.0
+
+### Patch Changes
+
+- Updated dependencies [c265721]
+  - @paprika/popover@1.2.0-next.0
+  - @paprika/confirmation@1.0.12-next.0
+
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [12938b6]
+  - @paprika/popover@1.1.0
+  - @paprika/confirmation@1.0.11
+
+## 1.0.13-next.0
+
+### Patch Changes
+
+- Updated dependencies [12938b6]
+  - @paprika/popover@1.1.0-next.0
+  - @paprika/confirmation@1.0.11-next.0
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).