+++ packages/Popover/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.2.0-next.0
+
+### Minor Changes
+
+- c265721: Fixed aria tags on popover
+
+## 1.1.0
+
+### Minor Changes
+
+- 12938b6: Added aria-haspopup to trigger
+
+## 1.1.0-next.0
+
+### Minor Changes
+
+- 12938b6: Added aria-haspopup to trigger
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).