+++ packages/Search/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @paprika/search
 
+## 1.0.9-next.0
+
+### Patch Changes
+
+- @paprika/list-box@3.0.2-next.0
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [0b3b614]
+  - @paprika/list-box@3.0.1
+
+## 1.0.8-next.0
+
+### Patch Changes
+
+- @paprika/list-box@3.0.1-next.0
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [47607db]
+  - @paprika/list-box@3.0.0
+
+## 1.0.7-next.0
+
+### Patch Changes
+
+- Updated dependencies [47607db]
+  - @paprika/list-box@3.0.0-next.0
+
 ## 1.0.6
 
 ### Patch Changes