+++ packages/Table/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.0.12-next.0
+
+### Patch Changes
+
+- @paprika/overflow-menu@1.0.14-next.0
+
+## 1.0.11
+
+### Patch Changes
+
+- @paprika/overflow-menu@1.0.13
+
+## 1.0.11-next.0
+
+### Patch Changes
+
+- @paprika/overflow-menu@1.0.13-next.0
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).