+++ packages/Toast/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## 1.1.0
+
+### Minor Changes
+
+- 0b3b614: #### Added
+
+  - `renderDelay` prop.
+
+  #### Fixed
+
+  - Render delay is always implemented, regardless of `isPolite` value.
+
+  [@mikrotron](https://github.com/mikrotron)
+
+## 1.1.0-next.0
+
+### Minor Changes
+
+- 0b3b614: #### Added
+
+  - `renderDelay` prop.
+
+  #### Fixed
+
+  - Render delay is always implemented, regardless of `isPolite` value.
+
+  [@mikrotron](https://github.com/mikrotron)
+
 ## [Unreleased]
 
 ## [] - 2020-07-29